+++ src/components/company/index.ts
@@ -1,5 +1,2 @@
 export { CompanyAddForm } from './CompanyAddForm';
 export { CompanyDomainStatus } from './CompanyDomainStatus';
-export { CompanyAddFormWrapper } from './CompanyAddFormWrapper';
-export { CompanyAddDetailsForm } from './CompanyAddDetailsForm';
-export { CompanyAddSettingsForm } from './CompanyAddSettingsForm';
